==> source/3.1-covariate-R.ipynb <==
### This code is for converting gene names in covariate file
### Auther Yiyun
rm(list=ls()) 
setwd('./')

library(biomaRt)
human = useMart("ensembl", dataset = "hsapiens_gene_ensembl", host= "grch37.ensembl.org")

###get gene list
dir_ref = '../data/proc_refs'
gene.list = read.csv(file.path(dir_ref,'gene_name_list_102121.txt'), header = FALSE)

res = getBM(attributes = c("external_gene_name","external_synonym"),
            filters = "external_gene_name",
            values = gene.list,
            mart = human)

### Write output
write.table(res, file.path(dir_ref,'gene.name.external.txt'), append = FALSE, sep = "\t", dec = "\t",
            row.names = FALSE, col.names = c("original","synonym"), quote = FALSE)
