==> source/3.2-covariate.ipynb <==
### This script re-annotate covariate file
# Author: Yiyun

import pandas as pd
import os, pickle, mmap
from multiprocessing import Pool

***
**re-annotate covariate file**

import math,pickle,os
import pandas as pd

### Read files
dir_ref = '../data/proc_refs'
dir_mutsig_in = '../data/MutSigCVsyn_inputs'

lgene = pickle.load(open(os.path.join(dir_ref,'gene_name_list_102121.pkl'),'rb'))
df_covar = pd.read_csv(os.path.join(dir_mutsig_in, 'gene.covariates.txt'),sep = '\t',index_col = 0)
df_covar_ext = pd.read_csv(os.path.join(dir_ref,'./gene.name.external.txt'), sep = '\t')
df_covar_ext = df_covar_ext.dropna()

### Drop na, drop duplicated genes in df_covar_ext
#1. drop if the synonyms is the same with original name
drop_idx = df_covar_ext[df_covar_ext.apply(pd.Series.nunique, axis=1) == 1].index.tolist()
idx_all = df_covar_ext.index.tolist()
keep_idx = [i for i in idx_all if i not in drop_idx]
df_covar_ext = df_covar_ext.loc[keep_idx,:]
# Drop if the synonym names are the same
df_covar_ext = df_covar_ext.drop_duplicates('synonym')
df_covar_ext = df_covar_ext.set_index('original')

lsub_name_all = {}
for genes in lgene:
    if genes not in df_covar.index:   # If the gene is not in covariate file
        print(genes)
        # find the gene in the external dataframe
        try:
            ext_names = df_covar_ext.loc[genes,'synonym']
            lsub_name = []
            # If the synonym name is in covariate file
            if isinstance(ext_names, str):
                if ext_names in df_covar.index.tolist():
                    lsub_name.append(ext_names)
            elif len(ext_names)> 0:
                for name in ext_names:
                    if name in df_covar.index.tolist():
                        lsub_name.append(name)
            lsub_name_all[genes] = lsub_name
        except KeyError:
            pass # some genes is not in the ext file

'SLAIN2' in lsub_name_all

n

for k,v in lsub_name_all.copy().items():
    # if 2 synonymous name is in the covariate file, means that one of them is not the gene we are referring to. 
    if len(v)>1: del lsub_name_all[k]
    # delete the one that is empty
    if len(v)==0: del lsub_name_all[k]

dict_sub = {v[0]: k for k, v in lsub_name_all.items()}

len(dict_sub) # number of genes to convert

df_covar = df_covar.rename(index=dict_sub)

# df_covar.to_csv(os.path.join(dir_mutsig_in,'gene.covariates.new.txt'), sep = '\t')
